encryption packaged renamed: secret.
