+++ migration/v6/model/core.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/google/uuid"
 	liberr "github.com/jortel/go-utils/error"
-	"github.com/konveyor/tackle2-hub/encryption"
+	"github.com/konveyor/tackle2-hub/secret"
 	"gorm.io/gorm"
 )
 
@@ -289,7 +289,7 @@ type Identity struct {
 // have changed and need to be (re)encrypted.
 func (r *Identity) Encrypt(ref *Identity) (err error) {
 	passphrase := Settings.Encryption.Passphrase
-	aes := encryption.AESCFB{}
+	aes := secret.AESCFB{}
 	aes.Use(passphrase)
 	if r.Password != ref.Password {
 		if r.Password != "" {
@@ -324,7 +324,7 @@ func (r *Identity) Encrypt(ref *Identity) (err error) {
 // Decrypt sensitive fields.
 func (r *Identity) Decrypt() (err error) {
 	passphrase := Settings.Encryption.Passphrase
-	aes := encryption.AESCFB{}
+	aes := secret.AESCFB{}
 	aes.Use(passphrase)
 	if r.Password != "" {
 		r.Password, err = aes.Decrypt(r.Password)
